+++ src/app.ts
@@ -1,11 +1,8 @@
 import {Server} from "./server/server";
 import express from "express";
-import MysqlHandler from "./server/database/mysqlHandler";
-import readline from 'readline'
-import AddFakeDataToDB from "./server/database/fakeData/addFakeDataToDB";
-const app = express();
 
 // --- Config ---
+const app = express();
 const port = 8080;
 const mySQLConnectionConfig = {
     host     : "localhost",
@@ -14,28 +11,5 @@ const mySQLConnectionConfig = {
     database : "timemanagerdatabase"
 }
 
-// Startup Node Server
-const server = new Server(app);
-server.start(port);
-
-//Create Mysql connection
-export const mysql: MysqlHandler = new MysqlHandler(mySQLConnectionConfig);
-
-//Console input
-const inquirer = readline.createInterface({
-    input: process.stdin,
-    output: process.stdout
-});
-
-inquirer.question("", input => {
-    if (input === "stop") {
-        inquirer.close();
-    }
-});
-
-inquirer.on("close", function() {
-    server.stop();
-    mysql.destroyConnection();
-    console.log("[Process] Everything shutdown, ending process");
-    process.exit(0);
-});
+// Startup Server
+export const server: Server = new Server(app, mySQLConnectionConfig, port);
